now we compute the occurrences also of object properties

+++ scripts/run-summarization.sh
@@ -347,6 +347,14 @@ echo "---Start: MinType, CountConcepts and CountDataTypes---"
 	    exit 1
 	fi
 
+	eval ${dbgCmd}""$JAVA_HOME/bin/java -Xms256m -Xmx16000m -cp ontology_summarization.jar it.unimib.disco.summarization.output.ProcessObjectRelationAssertions "${orgDatasetFile}" "$ResultsDirectory/patterns/"
+
+	if [ $? -ne 0 ]
+	then
+	    echo "App Failed during run"
+	    exit 1
+	fi
+
 	#Rimuovo i file dei tipi utilizzati per il calcolo, non più utili
 	rm -f ${orgDatasetFile}/*_types.nt
